ComposeWindow is the only entity that deals with passing events to ComposeScene
